requirements.txt
torch
ipython
ipdb
p_tqdm
sklearn
scipy
numpy
tqdm
mock
gitpython
